个人助手/requirements.txt
python-dotenv==1.0.0
SpeechRecognition==3.10.0
pyttsx3==2.90
langchain==0.1.0
langchain-openai==0.0.2
requests==2.31.0
pyautogui==0.9.54
Pillow==10.1.0
opencv-python==4.8.1.78
PyMuPDF==1.23.7
python-docx==1.0.1
python-pptx==0.6.21
tkinterdnd2==0.3.0 